Throw on JSON.stringify of an Integer
To future-proof against possible ways we could allow Integer in JSON,
this patch bans them for now. This option contrasts with Symbols,
which show up as missing in JSON object serialization. It's expected
that it will take longer to get Integer into a version of JSON than
it will to get Integer into browsers in memory, and that we don't
want to block the second on the first.

Relates to #24, but leaving that bug open to discuss the future
beyond the initial throwing plan.