Normative: Improve specification of conversion from String to BigInt
I had previously misunderstood the grammar used by this conversion;
there was a dead code path for empty strings which failed to parse,
whereas empty strings are converted to 0.

This patch cleans that up and uses NaN as an error sentinel, rather
than other tokens. It preserves the semantics of `==`, and the semantics
of `BigInt("")` are clarified to return `0n`.

Closes #66
Closes #64
Closes #60