Normative: Disallow BigInts as property keys
Previously, BigInts would behave like Numbers when used as property
keys, e.g., to index an array: They are converted to a String. So,
you could write code like this:

```js
  let arr = [];
  arr[0n] = 1;
  arr[1n] = 2;
  console.log(arr);  // 1,2
```

The behavior here falls out "naturally", but it may contribute to some
mistaken intuitions: You cannot use BigInts in other contexts which
require an Array index, such as Array methods which take an index
parameter (and do ToUint32), but users may expect to be able to.

Users may also expect the BigInt Array index path to have good
performance, even though it's not an encouraged strategy; this could
add unnecessary implementation burden.

This patch prohibits BigInts as property keys by throwing a TypeError
when they are used in that context.

Closes #58