Normative: Eliminate double-calling toJSON on BigInts in JSON.serialize
Previously, if toJSON on a BigInt wrapper returned a BigInt or
BigInt wrapper, then toJSON would be called a second time. Additionally,
the Get operation was used incorrectly.

This patch fixes the use of toJSON on BigInts in JSON.serialize to call
only once, before the replacer, at the same time as for objects. Later,
if a BigInt or wrapper is the result after the replacer (if provided),
a TypeError is thrown. GetV is used rather than Get, as is appropriate
for a primitive.

Closes #61