Normative: Remove == Boolean changes
Previously, Abstract Equality Comparison had been updated for
Boolean arguments in order to permit comparisons such as `0n == false`
back when comparisons like `0 == 0n` would still throw. Now, with
the latter comparison not throwing, the update is no longer necessary.
Further, @anba points out that the changes here are observable even
when BigInts are not present, and therefore present a compatibility
risk. This patch removes all changes for == on Booleans.
Addresses part of #79