Normative: Ban BigInt in LiteralPropertyName
This patch refactors the BigInt grammar to form a separate literal
production from NumericLiteral. In doing so, it removes BigInt
literals from LiteralPropertyName. For example, the following
would be a syntax error with this patch, and valid without it:
  ({ 5n: 1 })

This change is made in conjunction with the proposal to ban BigInts
as property keys, in #68. Even if we don't end up banning BigInts
as property keys, it still makes sense to not add them to the literal
grammar, as they will cast to a form which removes the `n`, making
them entirely useless and solely a source of confusion and curiosity.