Normative: Throw a TypeError earlier from TypedArrayCreate
In [this comment][comment], @jakobkummerow pointed out that the
BigInt TypedArray integration creates new cases which previously
did not throw exceptions and now do. The most tricky of those cases,
as far as I can see, is in methods like %TypedArray%.prototype.filter,
which call TypedArraySpeciesCreate and expect to get a result which
they can copy values from previous TypedArrays to. If a TypedArray
subclass overrides `@@species`, this process could create a
BigInt64Array from a Uint8Array subclass instance, for example.
However, a TypeError would happen as soon as you copy an entry
from one TypedArray to the other, as casts between Number and
BigInt need to be explicit.

This patch prohibits `@@species` of a TypedArray subclass from
pointing between a Big* TypedArray and a TypedArray whose
elements are represented by Numbers. In several of the methods
on %TypedArray%.prototype, an exception would be thrown shortly
afterwards if `@@species` were set in such a way. This patch
simply lifts that exception to a consistent place. It's hoped
that the change should make debugging as well as implementing
BigInt/TypedArray interaction easier.

[comment]: https://github.com/tc39/proposal-bigint/issues/102#issuecomment-351595780