Editorial: Call ToBigInt, not ToBigInt64, for casts
I don't believe there's an observable outcome, but in a couple cases
where the cast was recently changed to not call ToNumber in BigInt
cases, there was accidentally ToBigInt64 written. This patch calls
ToBigInt, which is analogous to the parallel ToNumber calls; there
will be a later call which will restrict it to the appropriate range.