Normative: Allow BigInt<->Number casts for TypedArray set
This patch explores a suggestion from @jakobkummerow in
https://github.com/tc39/proposal-bigint/issues/102#issuecomment-351868868

The logic of this change is that TypedArrays are already explicit in
their types, and it would be redundant to force users to insert additional
casts.

Namely, when writing into a TypedArray, there is a more flexible
cast operation used: If a BigInt is written into an existing TypedArray
type, it is implicitly cast to the appropriately sized type. Similarly,
a Number is implicitly cast to a 64-bit integer when writing it into
a BigInt64Array/BigUint64Array. Several TypedArray methods and functions
are updated to propagate the flexibility to them, including Atomics
and DataView functions.

This patch is not ready to merge until we have more discussion about
whether such a change is desirable.