Normative: Restore the IsSafeInteger check in BigInt()
When converting a Number to a BigInt in the BigInt constructor,
the Number is checked for whether it is a "safe integer". If it is
not, an exception is thrown. This logic was decided on in the
conclusion to https://github.com/tc39/proposal-bigint/issues/15 .

This typo was made in a previous patch due to a refactoring to
share a common path, which wasn't undone when that other path
removed the IsSafeInteger check (over the evolution of the PR).