Normative: Disallow tearing from atomic BigInt64 ops
The BigInt proposal adds atomic operations on BigInt64Array and
BigUint64Array. Nonatomic reads and writes on these TypedArrays
may or may not "tear", whereas atomic operations are expected to
not tear. Previously, the specification permitted tearing in all
cases; this patch clarifies that, when the operation is atomic,
then tearing does not occur.