Normative: Permit BigInt(large Number)
Previously, the BigInt constructor used IsSafeInteger to check if a Number
is safe to cast to a BigInt without losing information. This check is
unnecessarily conservative--it rules out large integer-valued Numbers.
This patch switches to just check that the Number is an integer value.

This patch specifies the resolution of the discussion in the March
2018 TC39 meeting.

Closes #132