Normative: Cast a String to a BigInt from <
For a comparison like 1n < "2", the previous semantics were to cast
"2" to 2, and then compare 1n and 2. The new semantics, in
this patch, is to cast "2" to 2n, and compare as BigInts. The
difference comes up most prominently for some large numbers, where
precision would previously be lost. Any comparison including a string
which is a non-integer will return false, to signify that the parsing
failed.

This patch specifies the semantics agreed on in the March 2018 TC39
meeting.

Closes #133